+++ src/main/java/com/solank/fxglgames/sg/components/weapons/SmallGun.java
@@ -44,7 +44,7 @@ public class SmallGun extends AbstractWeapon {
     }
 
     @Override
-    public double getDamage() {
+    public int getDamage() {
         return 10;
     }
